Allow macOS builds and bundling
SDLMain.h and SDLMain.m come from the SDL 1.2 distribution and are
required to build for macOS; other platforms should ignore them.
However, SDLMain.m has been modified so that the current working
directory is set to the .app Resources folder. This makes asset loading
code to work as expected.

Info.plist is metadata that holds the bundle together.

sdlsnake.iconset is the icon at various resolutions. I took the snake
head and scaled it without filtering. The bundle requires a .icns file,
which you can generate with:

   iconutil -c icns sdlsnake.iconset

To bundle as a .app:

 1. mkdir -p sdlsnake.app/Contents/MacOS
 2. mkdir -p sdlsnake.app/Contents/Frameworks
 3. mkdir -p sdlsnake.app/Contents/Resources
 4. Copy Info.plist into Contents/
 5. Copy the built binary into the MacOS/ folder
 6. Copy SDL.framework and SDL_mixer.framework into Frameworks/
 7. Copy the images/ and sounds/ folders into Resources/
 8. iconutil -c icns sdlsnake.iconset
 9. Copy sdlsnake.icns into Resources/

Resources:

  * https://stackoverflow.com/questions/6337787/how-can-i-set-the-icon-for-a-mac-application-in-xcode
  * http://meandmark.com/blog/2009/12/sdl-tips-for-mac-os-x/